Updated solution to use two priority queues (min heap backed)

+++ LeetCodeTests/Hard/4 Median of Two Sorted Arrays/ExtendedCases.cs
@@ -4,10 +4,10 @@ namespace LeetCodeTests.Hard._4_Median_of_Two_Sorted_Arrays;
 
 public class ExtendedCases {
     [Theory]
-    [InlineData(new[] { 1d }, new[] { 2d, 3d, 4d }, 2.5)]
-    [InlineData(new[] { 2d, 3d, 4d }, new[] { 1d }, 2.5)]
-    [InlineData(new[] { 3d, 4d }, new[] { 1d, 2d }, 2.5)]
-    public void NeverGetToMedianBeforeEndOfSpace(double[] num1, double[] num2, double expected) {
+    [InlineData(new[] { 1 }, new[] { 2, 3, 4 }, 2.5)]
+    [InlineData(new[] { 2, 3, 4 }, new[] { 1 }, 2.5)]
+    [InlineData(new[] { 3, 4 }, new[] { 1, 2 }, 2.5)]
+    public void NeverGetToMedianBeforeEndOfSpace(int[] num1, int[] num2, double expected) {
         Solution sut = new();
 
         var result = sut.FindMedianSortedArrays(num1, num2);
